pyproject.toml
[tool.poetry]
name = "homeworkdrf"
version = "0.1.0"
description = ""
authors = ["Nurlan-Grachyov <nurlan.grachyov@mail.ru>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
django = "^5.1.7"
djangorestframework = "^3.15.2"
pillow = "^11.1.0"
psycopg2-binary = "^2.9.10"
python-dotenv = "^1.0.1"
django-phonenumber-field = {extras = ["phonenumbers"], version = "^8.0.0"}
black = "^25.1.0"
flake8 = "^7.1.2"
isort = "^6.0.1"
django-filter = "^25.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
